Repurpose the repository to demonstrate the transformer
Since the issue that was being demonstrated in this repository and was found due to the fact that it was blocking the linked repository has been fixed now, I am repurposing the repository to demonstrate a TypeScript transformer enabled by the fact that the issue is no more. A more sophisticated version of the transformer will later be implemented in the linked repository which is now no longer blocked as a result of the issue being fixed.